// hivemq-edge-frontend/src/api/__generated__/models/SchemaReference.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * A schema reference is a unique identifier for a schema.
 */
export type SchemaReference = {
    /**
     * The identifier of the schema.
     */
    schemaId: string;
    /**
     * The version of the schema. The value "latest" may be used to always refer to the latest schema.
     */
    version: string;
};